telegram_bot/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
httpx==0.27.2
openai==1.54.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1
